trust-label/apps/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
openai==1.6.1
anthropic==0.8.1
langchain==0.0.352
chromadb==0.4.22
sentence-transformers==2.2.2
boto3==1.34.11
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9